Website/example fixes

+++ examples/getting-started/external-context/app.js
@@ -16,6 +16,11 @@ export default class AppAnimationLoop extends MiniAnimationLoop {
   start(props) {
     const canvas = this._getCanvas(props);
 
+    this.isDemoSupported = Boolean(canvas.getContext('webgl'));
+    if (!this.isDemoSupported) {
+      return;
+    }
+
     const gl = instrumentGLContext(canvas.getContext('webgl'));
     gl.clearColor(0, 0, 0, 1);
 
@@ -73,10 +78,18 @@ export default class AppAnimationLoop extends MiniAnimationLoop {
   }
 
   stop() {
+    if (!this.isDemoSupported) {
+      return;
+    }
+
     cancelAnimationFrame(this.resources.rafHandle);
   }
 
   delete() {
+    if (!this.isDemoSupported) {
+      return;
+    }
+
     const {positionBuffer, colorBuffer, model} = this.resources;
     positionBuffer.delete();
     colorBuffer.delete();
